Revert "added updates for minimal working example of probe and method hooking with user-init block that can be executed with armory run - requires necessary custom files and config not included in this repo"
This reverts commit c3570d51a3aaea436b96f402de5d987f8d28f358.

Reverting this commit based on discussion with Lucas and the need to focus on Probe examples that work under the assumption that the user is actively changing code to add Probes for a model or attack, as opposed to hooking a Probe to an existing model or attack without modifying the respective code directly - aside from updating the docs (I think at this point I'll rename the existing Probe.md as a different use case and add another doc just for the use case that I dicussed with Lucas), this will not change the fact that the user needs to use the user-init block (I may need to do a pull since the user-init implementation doesn't exist here yet - I've definitely seen it in the repo) with a Meter defined for the Probes that will be added, and it will also not change the need for perhaps a new Meter class that will export probe variables to a pickle file depending on what the user is trying to save (e.g. patch data or image with patch data) rather than a json file. There is also the question of how to deal with size restrictions placed on a Meter (unclear where the restrictions are actually defined - I only know that it exists because of armory logs and the fact that the variable doesn't get saved to the json file when I see size restriction logs).

An additional side note regarding the user-init block for the case that we do decide to add capabilities for Probe hooking without modifying the model or attack code: The way I got it to work in this commit (which will be reverted) is to load the user-init block at the end of all other loads (e.g. model, attack, dataset, metrics etc.) because an instance of a model or attack had to be available to hook to, which I accomplished by passing the Scenario (which contains the model and attack as an attribute) as an argument to a user init function. The existing user-init feature (again not in this branch currently, so will require a git pull and merge), assumes the user-init load happens prior to any other loads, which means none of the instances will be available just yet. Hopefully there will be no issue with this when defining a Meter to accompany any Probe the user may have added to a model or attack code, since the Meter can exist separately from a Probe instance that the Meter will get updates from.